Enhance push notifications and cleanup debug logs
- Integrate flutter_local_notifications for foreground display
- Add listener for foreground messages and invalidate relevant providers
- Implement deep navigation via GoRouter based on notification payload
- Fix main.dart initialization logic
- Remove debug print statements across the codebase
- General code cleanup and lint fixes

+++ lib/core/data/mappers/review_mapper.dart
@@ -39,7 +39,7 @@ class ReviewMapper {
       orderId: api.saleOrder?.toString() ?? '',
       reviewerId: api.reviewer?.id?.toString() ?? '',
       reviewedUserId: api.reviewedAccount?.id?.toString() ?? '',
-      rating: Rating.fromValue(api.rate),
+      rating: Rating.fromValue(api.rate.toDouble()),
       comment: api.message,
       timestamp: api.createdAt != null
           ? DateTime.tryParse(api.createdAt!) ?? DateTime.now()
